Seperate fetcher from getStitchedSchema
