+++ packages/federation/package.json
@@ -57,6 +57,7 @@
     "@graphql-tools/stitch": "^9.2.9",
     "@graphql-tools/utils": "^10.2.1",
     "@graphql-tools/wrap": "^10.0.3",
+    "@whatwg-node/fetch": "^0.9.17",
     "tslib": "^2.4.0",
     "value-or-promise": "^1.0.12"
   },
