+++ packages/federation/src/gateway.ts
@@ -232,136 +232,3 @@ export const federationSubschemaTransformer: SubschemaConfigTransform =
       },
     };
   };
-
-export type GetStitchedSchemaFromManagedFederationOpts = {
-  /**
-   * The graph ID of the managed federation graph.
-   */
-  graphId: string;
-  /**
-   * The API key to use to authenticate with the managed federation up link.
-   * It needs at least the `service:read` permission.
-   */
-  apiKey: string;
-  /**
-   * The URL of the managed federation up link.
-   * Default: 'https://uplink.api.apollographql.com/' (Apollo's managed federation up link on GCP)
-   * Alternative: 'https://aws.uplink.api.apollographql.com/' (Apollo's managed federation up link on AWS)
-   */
-  upLink?: string;
-  /**
-   * The ID of the last fetched supergraph.
-   * If provided, a supergraph is returned only if the managed supergraph have changed.
-   */
-  lastSeenId?: string;
-  /**
-   * The fetch implementation to use.
-   * Default: global.fetch
-   */
-  fetch?: typeof fetch;
-};
-
-export type RouterConfig = {
-  supergraphSdl: string;
-  minDelaySeconds: number;
-  id: string;
-};
-
-export type Unchanged = {
-  minDelaySeconds?: never;
-  id: string;
-};
-
-export type FetchError = {
-  error: { code: string; message: string };
-  minDelaySeconds: never;
-};
-
-/**
- * Fetches the supergraph SDL from a managed federation GraphOS up link.
- * @param options
- * @returns An object with the supergraph SDL when possible. It also includes metadata to handle polling and retry logic.
- *          If `lastSeenId` is provided and the supergraph has not changed, `supergraphSdl` is not present.
- */
-export async function getStitchedSchemaFromManagedFederation(
-  options: GetStitchedSchemaFromManagedFederationOpts,
-) {
-  const {
-    upLink = 'https://uplink.api.apollographql.com/',
-    fetch = global.fetch,
-    ...variables
-  } = options;
-
-  const response = await fetch(upLink, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      query: /* GraphQL */ `
-        query ($apiKey: String!, $graphId: String!, $lastSeenId: ID) {
-          routerConfig(ref: $graphId, apiKey: $apiKey, ifAfterId: $lastSeenId) {
-            __typename
-            ... on FetchError {
-              code
-              message
-              minDelaySeconds
-            }
-            ... on Unchanged {
-              id
-              minDelaySeconds
-            }
-            ... on RouterConfigResult {
-              id
-              supergraphSDL
-              minDelaySeconds
-            }
-          }
-        }
-      `,
-      variables,
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error(
-      `Failed to fetch supergraph SDL from managed federation up link '${upLink}': [${response.status} ${response.statusText}] ${await response.text()}`,
-    );
-  }
-
-  const responseBody = await response.text();
-  let result: ExecutionResult;
-  try {
-    result = JSON.parse(responseBody);
-  } catch (err) {
-    throw new Error(
-      `Failed to parse response from managed federation up link '${upLink}': ${(err as Error).message}\n\n${responseBody}`,
-    );
-  }
-
-  if (result.errors) {
-    const errors = result.errors.map(({ message }) => '\n' + message).join('');
-    throw new Error(
-      `Failed to fetch supergraph SDL from managed federation up link '${upLink}': ${errors}`,
-    );
-  }
-
-  const { routerConfig } = result.data;
-
-  if (routerConfig.__typename === 'FetchError') {
-    return {
-      error: { code: routerConfig.code, message: routerConfig.message },
-      minDelaySeconds: routerConfig.minDelaySeconds,
-    } as FetchError;
-  }
-
-  if (routerConfig.__typename === 'Unchanged') {
-    return { id: routerConfig.id, minDelaySeconds: routerConfig.minDelaySeconds } as Unchanged;
-  }
-
-  return {
-    supergraphSdl: routerConfig.supergraphSDL,
-    id: routerConfig.id,
-    minDelaySeconds: routerConfig.minDelaySeconds,
-  } as RouterConfig;
-}
